Revert "feat: RPA continuous feed for notify claim (CMC-1613) (#141)"
This reverts commit a626cb68f85fd7c3954ff71f855b850d19173312.
