+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -75,7 +75,6 @@ public enum CaseEvent {
     GENERATE_ACKNOWLEDGEMENT_OF_CLAIM(CAMUNDA),
     GENERATE_DIRECTIONS_QUESTIONNAIRE(CAMUNDA),
     NOTIFY_RPA_ON_CASE_HANDED_OFFLINE(CAMUNDA),
-    NOTIFY_RPA_ON_CONTINUOUS_FEED(CAMUNDA),
     RETRY_NOTIFY_RPA_ON_CASE_HANDED_OFFLINE(CAMUNDA),
     RESET_RPA_NOTIFICATION_BUSINESS_PROCESS(CAMUNDA),
