+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/robotics/NotifyRoboticsOnCaseHandedOfflineHandler.java
@@ -1,41 +1,52 @@
 package uk.gov.hmcts.reform.civil.handler.callback.camunda.robotics;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.networknt.schema.ValidationMessage;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
 import uk.gov.hmcts.reform.civil.callback.Callback;
+import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.robotics.RoboticsCaseData;
+import uk.gov.hmcts.reform.civil.service.Time;
 import uk.gov.hmcts.reform.civil.service.robotics.JsonSchemaValidationService;
 import uk.gov.hmcts.reform.civil.service.robotics.RoboticsNotificationService;
+import uk.gov.hmcts.reform.civil.service.robotics.exception.JsonSchemaValidationException;
+import uk.gov.hmcts.reform.civil.service.robotics.exception.RoboticsDataException;
 import uk.gov.hmcts.reform.civil.service.robotics.mapper.RoboticsDataMapper;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
+import static java.lang.String.format;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_RPA_ON_CASE_HANDED_OFFLINE;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.RETRY_NOTIFY_RPA_ON_CASE_HANDED_OFFLINE;
 
 @Service
-public class NotifyRoboticsOnCaseHandedOfflineHandler extends NotifyRoboticsHandler {
+@RequiredArgsConstructor
+public class NotifyRoboticsOnCaseHandedOfflineHandler extends CallbackHandler {
 
-    public NotifyRoboticsOnCaseHandedOfflineHandler(
-        RoboticsNotificationService roboticsNotificationService,
-        JsonSchemaValidationService jsonSchemaValidationService,
-        RoboticsDataMapper roboticsDataMapper
-    ) {
-        super(roboticsNotificationService, jsonSchemaValidationService, roboticsDataMapper);
-    }
-
-    private static final List<CaseEvent> EVENTS = List.of(
-        NOTIFY_RPA_ON_CASE_HANDED_OFFLINE,
-        RETRY_NOTIFY_RPA_ON_CASE_HANDED_OFFLINE
-    );
+    private static final List<CaseEvent> EVENTS = List.of(NOTIFY_RPA_ON_CASE_HANDED_OFFLINE,
+                                                          RETRY_NOTIFY_RPA_ON_CASE_HANDED_OFFLINE);
     public static final String TASK_ID = "NotifyRoboticsOnCaseHandedOffline";
+
+    private final RoboticsNotificationService roboticsNotificationService;
+    private final JsonSchemaValidationService jsonSchemaValidationService;
+    private final RoboticsDataMapper roboticsDataMapper;
+    private final Time time;
+    private final ObjectMapper mapper;
 
     @Override
     protected Map<String, Callback> callbacks() {
         return Map.of(
-            callbackKey(ABOUT_TO_SUBMIT), this::notifyRobotics
+            callbackKey(ABOUT_TO_SUBMIT), this::notifyRoboticsForCaseHandedOffline
         );
     }
 
@@ -47,5 +58,25 @@ public class NotifyRoboticsOnCaseHandedOfflineHandler extends NotifyRoboticsHand
     @Override
     public List<CaseEvent> handledEvents() {
         return EVENTS;
+    }
+
+    private CallbackResponse notifyRoboticsForCaseHandedOffline(CallbackParams callbackParams) {
+        CaseData caseData = callbackParams.getCaseData();
+
+        try {
+            RoboticsCaseData roboticsCaseData = roboticsDataMapper.toRoboticsCaseData(caseData);
+            Set<ValidationMessage> errors = jsonSchemaValidationService.validate(roboticsCaseData.toJsonString());
+            if (errors.isEmpty()) {
+                roboticsNotificationService.notifyRobotics(caseData);
+            } else {
+                throw new JsonSchemaValidationException(
+                    format("Invalid RPA Json payload for %s", caseData.getLegacyCaseReference()),
+                    errors
+                );
+            }
+        } catch (JsonProcessingException e) {
+            throw new RoboticsDataException(e.getMessage(), e);
+        }
+        return AboutToStartOrSubmitCallbackResponse.builder().build();
     }
 }
