+++ src/test/java/uk/gov/hmcts/reform/civil/sampledata/CaseDataMinEdgeCasesBuilder.java
@@ -15,7 +15,6 @@ import uk.gov.hmcts.reform.civil.model.SolicitorOrganisationDetails;
 import uk.gov.hmcts.reform.civil.model.StatementOfTruth;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 import static java.math.BigDecimal.TEN;
@@ -101,14 +100,6 @@ public class CaseDataMinEdgeCasesBuilder extends CaseDataBuilder {
         return this;
     }
 
-    public CaseDataMinEdgeCasesBuilder atStateClaimIssuedWithMinimalData() {
-        atStatePendingCaseIssuedWithMinimalData();
-        issueDate = CLAIM_ISSUED_DATE;
-        claimNotificationDeadline = NOTIFICATION_DEADLINE;
-        ccdState = CASE_ISSUED;
-        return this;
-    }
-
     public CaseDataMinEdgeCasesBuilder atStatePaymentSuccessfulWithMinimalData() {
         atStatePendingCaseIssuedWithMinimalData();
         claimIssuedPaymentDetails = PaymentDetails.builder()
@@ -120,7 +111,7 @@ public class CaseDataMinEdgeCasesBuilder extends CaseDataBuilder {
         return this;
     }
 
-    public CaseDataMinEdgeCasesBuilder atStateAwaitingCaseNotificationWithMinimalData() {
+    public CaseDataBuilder atStateAwaitingCaseNotificationWithMinimalData() {
         atStatePaymentSuccessfulWithMinimalData();
         ccdState = CASE_ISSUED;
         issueDate = CLAIM_ISSUED_DATE;
@@ -128,14 +119,6 @@ public class CaseDataMinEdgeCasesBuilder extends CaseDataBuilder {
         return this;
     }
 
-    public CaseDataMinEdgeCasesBuilder atStateClaimNotifiedWithMaximumData() {
-        atStateClaimIssuedWithMinimalData();
-        claimNotificationDate = LocalDate.now().atStartOfDay();
-        claimDetailsNotificationDeadline = DEADLINE;
-        ccdState = AWAITING_CASE_DETAILS_NOTIFICATION;
-        return this;
-    }
-    
     public CaseDataMinEdgeCasesBuilder atStateAwaitingCaseDetailsNotificationWithMinimalData() {
         atStateAwaitingCaseNotificationWithMinimalData();
         claimNotificationDate = LocalDateTime.now();
